Conformance tests refactoring
Some of the code in test/conformance/util.go was referring to code
inside test/conformance/blue_green.go which should ideally be part of
util. The change moves that code.
